pyproject.toml
[project]
name = "promptx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.3.8",
    "json-repair>=0.55.1",
    "jupyter>=1.1.1",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.1.7",
    "langfuse>=3.12.1",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.1",
]
